pyproject.toml
[project]
name = "promptproof"
version = "0.1.0"
description = "PromptProof - local LLM red-team-in-a-box"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.12",
  "pydantic>=2.7",
  "rich>=13.7",
  "jinja2>=3.1",
  "ruamel.yaml>=0.18",
  "httpx>=0.27",
  "rapidfuzz>=3.9",
  "regex>=2024.5"
]

[project.scripts]
promptproof = "promptproof.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["promptproof*"]
exclude = ["attacks*", "configs*", "data*", "logs*", "reports*", "scripts*"]
